mlocate: Add updatedb.conf and backport a patch
The conf file is needed to prevent indexing volatile directories, like
the contents of /proc/[pid] and things like /tmp and /var/tmp.
findutils has these as defaults, but mlocate does not.

This updatedb.conf is taken from fedora, written by the mlocate author.
https://pkgs.fedoraproject.org/cgit/mlocate.git/tree/updatedb.conf

And the patch should speak for its self.

Change-Id: If474b80a0f90e58776b75ad8c40b09b51f4cfa94
Reviewed-on: https://galileo.mailstation.de/gerrit/4792
Reviewed-by: Quentin Glidic <sardemff7+git@sardemff7.net>